--- a/packages/canvas-extract/test/CanvasExtract.tests.ts
+++ b/packages/canvas-extract/test/CanvasExtract.tests.ts
@@ -1,13 +1,13 @@
 import { CanvasExtract } from '@pixi/canvas-extract';
+import { CanvasGraphicsRenderer } from '@pixi/canvas-graphics';
 import { CanvasRenderer } from '@pixi/canvas-renderer';
 import { RenderTexture, Texture } from '@pixi/core';
+import { Graphics } from '@pixi/graphics';
 import { Sprite } from '@pixi/sprite';
 import '@pixi/canvas-display';
 
-describe('CanvasExtract', () =>
-{
-    it('should access extract on renderer', () =>
-    {
+describe('CanvasExtract', () => {
+    it('should access extract on renderer', () => {
         const renderer = new CanvasRenderer();
 
         expect(renderer.extract).toBeInstanceOf(CanvasExtract);
@@ -15,8 +15,111 @@ describe('CanvasExtract', () =>
         renderer.destroy();
     });
 
-    it('should extract an sprite', async () =>
-    {
+    it('should extract pixels from renderer correctly', async () => {
+        const renderer = new CanvasRenderer({ width: 2, height: 2 });
+
+        renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
+
+        const graphics = new Graphics()
+            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
+            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
+            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
+            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+        const extract = renderer.extract;
+
+        renderer.render(graphics);
+
+        const extractedPixels = extract.pixels();
+
+        expect(extractedPixels).toEqual(new Uint8ClampedArray([
+            255, 0, 0, 255, 0, 255, 0, 255,
+            0, 0, 255, 255, 255, 255, 0, 255
+        ]));
+
+        graphics.destroy();
+        renderer.destroy();
+    });
+
+    it('should extract canvas from renderer correctly', async () => {
+        const renderer = new CanvasRenderer({ width: 2, height: 2 });
+
+        renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
+
+        const graphics = new Graphics()
+            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
+            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
+            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
+            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+        const extract = renderer.extract;
+
+        renderer.render(graphics);
+
+        const canvas = extract.canvas();
+        const context = canvas.getContext('2d');
+        const imageData = context.getImageData(0, 0, 2, 2);
+
+        expect(imageData.data).toEqual(new Uint8ClampedArray([
+            255, 0, 0, 255, 0, 255, 0, 255,
+            0, 0, 255, 255, 255, 255, 0, 255
+        ]));
+
+        graphics.destroy();
+        renderer.destroy();
+    });
+
+    it('should extract pixels from render texture correctly', async () => {
+        const renderer = new CanvasRenderer({ width: 2, height: 2 });
+
+        renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
+
+        const graphics = new Graphics()
+            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
+            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
+            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
+            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+        const extract = renderer.extract;
+
+        renderer.render(graphics);
+
+        const extractedPixels = extract.pixels(graphics);
+
+        expect(extractedPixels).toEqual(new Uint8ClampedArray([
+            255, 0, 0, 255, 0, 255, 0, 255,
+            0, 0, 255, 255, 255, 255, 0, 255
+        ]));
+
+        graphics.destroy();
+        renderer.destroy();
+    });
+
+    it('should extract canvas from render texture correctly', async () => {
+        const renderer = new CanvasRenderer({ width: 2, height: 2 });
+
+        renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
+
+        const graphics = new Graphics()
+            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
+            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
+            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
+            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+        const extract = renderer.extract;
+
+        renderer.render(graphics);
+
+        const canvas = extract.canvas(graphics);
+        const context = canvas.getContext('2d');
+        const imageData = context.getImageData(0, 0, 2, 2);
+
+        expect(imageData.data).toEqual(new Uint8ClampedArray([
+            255, 0, 0, 255, 0, 255, 0, 255,
+            0, 0, 255, 255, 255, 255, 0, 255
+        ]));
+
+        graphics.destroy();
+        renderer.destroy();
+    });
+
+    it('should extract an sprite', async () => {
         const renderer = new CanvasRenderer();
         const sprite = new Sprite(Texture.WHITE);
         const extract = renderer.extract as CanvasExtract;
@@ -30,8 +133,7 @@ describe('CanvasExtract', () =>
         sprite.destroy();
     });
 
-    it('should extract with no arguments', async () =>
-    {
+    it('should extract with no arguments', async () => {
         const renderer = new CanvasRenderer();
         const extract = renderer.extract as CanvasExtract;
 
@@ -43,8 +145,7 @@ describe('CanvasExtract', () =>
         renderer.destroy();
     });
 
-    it('should extract a render texture', async () =>
-    {
+    it('should extract a render texture', async () => {
         const renderer = new CanvasRenderer();
         const extract = renderer.extract as CanvasExtract;
         const renderTexture = RenderTexture.create({ width: 10, height: 10 });