--- a/packages/canvas-extract/test/CanvasExtract.tests.ts
+++ b/packages/canvas-extract/test/CanvasExtract.tests.ts
@@ -6,8 +6,10 @@ import { Graphics } from '@pixi/graphics';
 import { Sprite } from '@pixi/sprite';
 import '@pixi/canvas-display';
 
-describe('CanvasExtract', () => {
-    it('should access extract on renderer', () => {
+describe('CanvasExtract', () =>
+{
+    it('should access extract on renderer', () =>
+    {
         const renderer = new CanvasRenderer();
 
         expect(renderer.extract).toBeInstanceOf(CanvasExtract);
@@ -15,16 +17,25 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract pixels from renderer correctly', async () => {
+    it('should extract pixels from renderer correctly', async () =>
+    {
         const renderer = new CanvasRenderer({ width: 2, height: 2 });
 
         renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
 
         const graphics = new Graphics()
-            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
-            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
-            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
-            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+            .beginFill(0xFF0000)
+            .drawRect(0, 0, 1, 1)
+            .endFill()
+            .beginFill(0x00FF00)
+            .drawRect(1, 0, 1, 1)
+            .endFill()
+            .beginFill(0x0000FF)
+            .drawRect(0, 1, 1, 1)
+            .endFill()
+            .beginFill(0xFFFF00)
+            .drawRect(1, 1, 1, 1)
+            .endFill();
         const extract = renderer.extract;
 
         renderer.render(graphics);
@@ -40,16 +51,25 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract canvas from renderer correctly', async () => {
+    it('should extract canvas from renderer correctly', async () =>
+    {
         const renderer = new CanvasRenderer({ width: 2, height: 2 });
 
         renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
 
         const graphics = new Graphics()
-            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
-            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
-            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
-            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+            .beginFill(0xFF0000)
+            .drawRect(0, 0, 1, 1)
+            .endFill()
+            .beginFill(0x00FF00)
+            .drawRect(1, 0, 1, 1)
+            .endFill()
+            .beginFill(0x0000FF)
+            .drawRect(0, 1, 1, 1)
+            .endFill()
+            .beginFill(0xFFFF00)
+            .drawRect(1, 1, 1, 1)
+            .endFill();
         const extract = renderer.extract;
 
         renderer.render(graphics);
@@ -67,16 +87,25 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract pixels from render texture correctly', async () => {
+    it('should extract pixels from render texture correctly', async () =>
+    {
         const renderer = new CanvasRenderer({ width: 2, height: 2 });
 
         renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
 
         const graphics = new Graphics()
-            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
-            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
-            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
-            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+            .beginFill(0xFF0000)
+            .drawRect(0, 0, 1, 1)
+            .endFill()
+            .beginFill(0x00FF00)
+            .drawRect(1, 0, 1, 1)
+            .endFill()
+            .beginFill(0x0000FF)
+            .drawRect(0, 1, 1, 1)
+            .endFill()
+            .beginFill(0xFFFF00)
+            .drawRect(1, 1, 1, 1)
+            .endFill();
         const extract = renderer.extract;
 
         renderer.render(graphics);
@@ -92,16 +121,25 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract canvas from render texture correctly', async () => {
+    it('should extract canvas from render texture correctly', async () =>
+    {
         const renderer = new CanvasRenderer({ width: 2, height: 2 });
 
         renderer.plugins.graphics = new CanvasGraphicsRenderer(renderer);
 
         const graphics = new Graphics()
-            .beginFill(0xFF0000).drawRect(0, 0, 1, 1).endFill()
-            .beginFill(0x00FF00).drawRect(1, 0, 1, 1).endFill()
-            .beginFill(0x0000FF).drawRect(0, 1, 1, 1).endFill()
-            .beginFill(0xFFFF00).drawRect(1, 1, 1, 1).endFill();
+            .beginFill(0xFF0000)
+            .drawRect(0, 0, 1, 1)
+            .endFill()
+            .beginFill(0x00FF00)
+            .drawRect(1, 0, 1, 1)
+            .endFill()
+            .beginFill(0x0000FF)
+            .drawRect(0, 1, 1, 1)
+            .endFill()
+            .beginFill(0xFFFF00)
+            .drawRect(1, 1, 1, 1)
+            .endFill();
         const extract = renderer.extract;
 
         renderer.render(graphics);
@@ -119,7 +157,8 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract an sprite', async () => {
+    it('should extract an sprite', async () =>
+    {
         const renderer = new CanvasRenderer();
         const sprite = new Sprite(Texture.WHITE);
         const extract = renderer.extract as CanvasExtract;
@@ -133,7 +172,8 @@ describe('CanvasExtract', () => {
         sprite.destroy();
     });
 
-    it('should extract with no arguments', async () => {
+    it('should extract with no arguments', async () =>
+    {
         const renderer = new CanvasRenderer();
         const extract = renderer.extract as CanvasExtract;
 
@@ -145,7 +185,8 @@ describe('CanvasExtract', () => {
         renderer.destroy();
     });
 
-    it('should extract a render texture', async () => {
+    it('should extract a render texture', async () =>
+    {
         const renderer = new CanvasRenderer();
         const extract = renderer.extract as CanvasExtract;
         const renderTexture = RenderTexture.create({ width: 10, height: 10 });