--- a/packages/extract/src/Extract.ts
+++ b/packages/extract/src/Extract.ts
@@ -86,7 +86,7 @@ export class Extract implements ISystem, IExtract
 
         if (canvas.toBlob !== undefined)
         {
-            return await new Promise<string>((resolve, reject) =>
+            return new Promise<string>((resolve, reject) =>
             {
                 canvas.toBlob((blob) =>
                 {
@@ -106,7 +106,7 @@ export class Extract implements ISystem, IExtract
         {
             const blob = await canvas.convertToBlob({ type: format, quality });
 
-            return await new Promise<string>((resolve, reject) =>
+            return new Promise<string>((resolve, reject) =>
             {
                 const reader = new FileReader();