##### Description of change

- 3d2d89c9d6ad1a6fe6bfb78fec5a2a4dc10d9d8f: `canvas.toBlob` instead of `toDataURL` blocks the main thread for a shorter duration.
- 4a2ff59114c9f015b9c96f7f1c04fd4e53c331c2: Improves performance (~50ms => ~42ms in my test). The `getImageData` is unnessecary.
- 4b2ce461ddd34465d27a138ed245c7eaf9d876ca: Improves performance even more (~42ms => ~34ms in my test).

##### Pre-Merge Checklist

- [x] Lint process passed (`npm run lint`)
- [x] Tests passed (`npm run test`)