@dev7355608 2023-04-03
I noticed my y-flipping was flawed: I forgot about the 4 bytes per pixel. There was no test that could have caught this error: so I added a bunch of tests that checks that looks that the extracted pixels/canvas.

I also replaced `postArrayDivide` with a much faster version (more that 3 times faster in my test).