@dev7355608 2023-04-03
I added test for `CanvasExtract` too. Notice that `Extract.pixels` from a renderer does not y-flip, while `CanvasExtract.pixels` returns y-flipped results. Is this a bug? Wouldn't it make more sense if `Extract.pixels` y-flipped?

Another difference that bothers me a little bit is that `CanvasExtract.pixels` returns `Uint8ClampedArray` and `Extract.pixels` returns `Uint8Array`.